backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
#insightface
jinja2==3.1.5
jiter==0.8.2
kiwisolver==1.4.8
linkedin-api==2.3.1
lxml==5.3.0
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.1



openai==1.82.1
opencv-python==4.10.0.84
orjson==3.10.13
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==24.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.4
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.18.0
pyparsing==3.2.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==21.10
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
ruff==0.8.4
safehttpx==0.1.6
scipy==1.14.1
scikit-learn==1.6.1
seaborn==0.13.2
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
stem==1.8.0
supervision==0.25.1
sympy==1.13.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.15.1
typing-extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.55
ultralytics-thop==2.0.13
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
werkzeug==3.1.3
wheel==0.44.0
whois==1.20240129.2
wsproto==1.2.0
yarl==1.18.3
spacy==3.8.5
fuzzywuzzy==0.18.0
uvicorn
python-Levenshtein